+++ backend/api/Program.cs
@@ -27,6 +27,7 @@ if (builder.Environment.IsProduction())
     builder.Services.AddNpgsqlDataSource(Utilities.ProperlyFormattedConnectionString);
 }
 builder.Services.AddAvatarBlobService();
+builder.Services.AddSingleton<TransactionCalculator>();
 builder.Services.AddSingleton<NotificationFacade>();
 
 builder.Services.AddSingleton<UserRepository>();
