+++ frontend/src/app/group/activity/activity.component.scss
@@ -19,5 +19,5 @@ ion-spinner {
 }
 
 .subpage {
-  height: 67vh;
+  height: 110vh;
 }
