--- frontend/src/app/group/balancecard/payee-transfor-card/payee-transfor-card.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { PayeeTransforCardComponent } from './payee-transfor-card.component';
-
-describe('PayeeTransforCardComponent', () => {
-  let component: PayeeTransforCardComponent;
-  let fixture: ComponentFixture<PayeeTransforCardComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PayeeTransforCardComponent ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(PayeeTransforCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
