--- frontend/src/app/group/balancecard/payer-transfor-card/payer-transfor-card.component.spec.ts
@@ -1,24 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { PayerTransforCardComponent } from './payer-transfor-card.component';
-
-describe('PayerTransforCardComponent', () => {
-  let component: PayerTransforCardComponent;
-  let fixture: ComponentFixture<PayerTransforCardComponent>;
-
-  beforeEach(waitForAsync(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PayerTransforCardComponent ],
-      imports: [IonicModule.forRoot()]
-    }).compileComponents();
-
-    fixture = TestBed.createComponent(PayerTransforCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  }));
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
